Convert pressure to altitude, support calibration of the standard pressure level

+++ src/com/vaslabs/sdc/sensors/SensorValue.java
@@ -2,19 +2,19 @@ package com.vaslabs.sdc.sensors;
 
 public abstract class SensorValue {
     
-    private double raw_value;
+    private float raw_value;
     private boolean isInitialised = false;
     
     public boolean hasBeenInitialised() {
         return isInitialised;
     }
     
-    public void setRawValue(double value) {
+    public void setRawValue(float value) {
         raw_value = value;
         isInitialised = true;
     }
 
-    public double getRawValue() {
+    public float getRawValue() {
         return raw_value;
     }
 }
